+++ oauth2_provider/generators.py
@@ -1,7 +1,5 @@
-from __future__ import absolute_import, unicode_literals
-
-from oauthlib.common import generate_client_id as oauthlib_generate_client_id
 from oauthlib.common import UNICODE_ASCII_CHARACTER_SET
+from oauthlib.common import generate_client_id as oauthlib_generate_client_id
 
 from .settings import oauth2_settings