+++ oauth2_provider/migrations/0003_auto_20160316_1503.py
@@ -1,20 +1,17 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
-from django.db import migrations, models
 from django.conf import settings
+from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('oauth2_provider', '0002_08_updates'),
+        ("oauth2_provider", "0002_08_updates"),
     ]
 
     operations = [
         migrations.AlterField(
-            model_name='application',
-            name='user',
-            field=models.ForeignKey(related_name='oauth2_provider_application', blank=True, to=settings.AUTH_USER_MODEL, null=True, on_delete=models.CASCADE),
+            model_name="application",
+            name="user",
+            field=models.ForeignKey(related_name="oauth2_provider_application", blank=True, to=settings.AUTH_USER_MODEL, null=True, on_delete=models.CASCADE),
         ),
     ]