+++ oauth2_provider/views/application.py
@@ -38,7 +38,7 @@ class ApplicationRegistration(LoginRequiredMixin, CreateView):
 
     def form_valid(self, form):
         form.instance.user = self.request.user
-        return super(ApplicationRegistration, self).form_valid(form)
+        return super().form_valid(form)
 
 
 class ApplicationDetail(ApplicationOwnerIsUserMixin, DetailView):