+++ tests/test_oauth2_validators.py
@@ -1,5 +1,5 @@
-import datetime
 import contextlib
+import datetime
 
 from django.contrib.auth import get_user_model
 from django.test import TransactionTestCase